refactor: split attribution stateful returns
